handle near bulk payment creation

+++ bootstrap/src/test/java/onlydust/com/marketplace/api/helper/AccountingHelper.java
@@ -203,15 +203,15 @@ public class AccountingHelper {
                 .account().id();
     }
 
-    public void allocate(SponsorId sponsorId, ProgramId programId, long amount, Currency.Id currencyId) {
+    public void allocate(SponsorId sponsorId, ProgramId programId, double amount, Currency.Id currencyId) {
         accountingFacadePort.allocate(sponsorId, programId, PositiveAmount.of(amount), currencyId);
     }
 
-    public void unallocate(ProgramId programId, SponsorId sponsorId, long amount, Currency.Id currencyId) {
+    public void unallocate(ProgramId programId, SponsorId sponsorId, double amount, Currency.Id currencyId) {
         accountingFacadePort.unallocate(programId, sponsorId, PositiveAmount.of(amount), currencyId);
     }
 
-    public void grant(ProgramId programId, ProjectId projectId, long amount, Currency.Id currencyId) {
+    public void grant(ProgramId programId, ProjectId projectId, double amount, Currency.Id currencyId) {
         accountingFacadePort.grant(programId, projectId, PositiveAmount.of(amount), currencyId);
     }
 
@@ -219,7 +219,7 @@ public class AccountingHelper {
         accountingFacadePort.pay(Set.of(rewardIds));
     }
 
-    public void ungrant(ProjectId projectId, ProgramId programId, long amount, Currency.Id currencyId) {
+    public void ungrant(ProjectId projectId, ProgramId programId, double amount, Currency.Id currencyId) {
         accountingFacadePort.ungrant(projectId, programId, PositiveAmount.of(amount), currencyId);
     }
