+++ bootstrap/src/test/java/onlydust/com/marketplace/api/helper/RewardHelper.java
@@ -23,7 +23,7 @@ public class RewardHelper {
     @Autowired
     private RewardStoragePort rewardStoragePort;
 
-    public RewardId create(ProjectId projectId, UserAuthHelper.AuthenticatedUser lead, GithubUserId recipientId, long amount, Currency.Id currencyId) {
+    public RewardId create(ProjectId projectId, UserAuthHelper.AuthenticatedUser lead, GithubUserId recipientId, double amount, Currency.Id currencyId) {
         return create(projectId, lead, recipientId, amount, currencyId, List.of(RequestRewardCommand.Item.builder()
                 .id("1974448961")
                 .number(77L)
@@ -32,7 +32,7 @@ public class RewardHelper {
                 .build()));
     }
 
-    public RewardId create(ProjectId projectId, UserAuthHelper.AuthenticatedUser lead, GithubUserId recipientId, long amount, Currency.Id currencyId,
+    public RewardId create(ProjectId projectId, UserAuthHelper.AuthenticatedUser lead, GithubUserId recipientId, double amount, Currency.Id currencyId,
                            List<RequestRewardCommand.Item> rewardItems) {
         return rewardFacadePort.createReward(UserId.of(lead.user().getId()),
                 RequestRewardCommand.builder()
